Move JuliaOptions etc. to julia.options
